fix: confliction when onnx convert
